Implement a discrete user dropdown menu for contextual operations. In addition, the user action dropdown will now update to accomodate permission changes live.
